fix bugs in tasks
