+++ main.py
@@ -2,8 +2,9 @@ from machine import Timer
 import time
 import gc
 import network
-import epaper2in9
-
+from machine import SPI
+import framebuf
+import utime
 import urequests as requests
 import json
 import gc
@@ -12,6 +13,38 @@ from machine import Pin
 import ntptime
 import time
 t = Timer()
+
+# Display resolution
+EPD_WIDTH       = 128
+EPD_HEIGHT      = 296
+
+RST_PIN         = 12
+DC_PIN          = 8
+CS_PIN          = 9
+BUSY_PIN        = 13
+
+WF_PARTIAL_2IN9 = [
+    0x0,0x40,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x80,0x80,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x40,0x40,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x80,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0A,0x0,0x0,0x0,0x0,0x0,0x1,  
+    0x1,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x1,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x0,0x0,0x0,0x0,0x0,0x0,0x0,
+    0x22,0x22,0x22,0x22,0x22,0x22,0x0,0x0,0x0,
+    0x22,0x17,0x41,0xB0,0x32,0x36,
+]
+
 
 class LoggerService():
     def info(self, message):
@@ -32,24 +65,23 @@ class LoggerService():
     
 
 class EpaperService():
-    def __init__(self):
-        self.epd = EPD_2in9_Landscape()
+    def __init__(self, els):
+        self.els = els
         
     def write(self, text, row):
         rowHeight = 30
         loc = rowHeight * row
-        
-        epd.Clear(0xff)
-    
-        epd.fill(0xff)
-        epd.text(text, 5, 10, 0x00)
+        #els.Clear(0xff)
+        #els.fill(0xff)
+        #els.text(text, 5, 10, 0x00)
 
 class TasksService():
-    def __init__(self, settingsService, loggerService):
+    def __init__(self, settingsService, loggerService, epaperService):
         self.settingsService = settingsService
         self.loggerService = loggerService
         self.token = settingsService.get("token")
         self.tasksUrl = settingsService.get("tasksUrl")
+        self.epaperService = epaperService
 
     def getTasks(self):
         header_data = { 
@@ -66,10 +98,15 @@ class TasksService():
 
     def displayTasks(self):  
         items = self.getTasks()
+        i = 1
         for item in items:
-            self.loggerService.info("Task:" + item["content"])
+            self.loggerService.trace("Task:" + item["content"])
             rowText = item["content"]
-            #epaperService.write(rowText, 1)
+            if self.epaperService is None:
+                self.loggerService.warn("Task: No epaper")
+            else:
+                self.epaperService.write(rowText, i)
+            i = i +1
 
     def getFirstTask(self, val): 
         for value in val:
@@ -101,7 +138,7 @@ class WifiService():
             self.loggerService.info("WiFi: Connected")
             internalLED.on()
         else:
-            self.loggerService.info("WiFi: Failed to connect")
+            self.loggerService.error("WiFi: Failed to connect")
             internalLED.off()
 
 
@@ -139,19 +176,434 @@ class App():
         self.tasksService = tasksService
 
     def main(self):
-        self.loggerService.info("Main: Start main loop")
+        self.loggerService.trace("Main: Start main loop")
         wlan = self.wifiService.connect() 
         time.sleep(3)   
         self.tasksService.displayTasks()
         time.sleep(5)
         wlan = self.wifiService.disconnect()
-        self.loggerService.info("Main: End main loop")
+        self.loggerService.trace("Main: End main loop")
         time.sleep(10)
 
 
+
+
+
+class EPD_2in9_Portrait(framebuf.FrameBuffer):
+    def __init__(self):
+        self.reset_pin = Pin(RST_PIN, Pin.OUT)
+        
+        self.busy_pin = Pin(BUSY_PIN, Pin.IN, Pin.PULL_UP)
+        self.cs_pin = Pin(CS_PIN, Pin.OUT)
+        self.width = EPD_WIDTH
+        self.height = EPD_HEIGHT
+        
+        self.lut = WF_PARTIAL_2IN9
+        
+        self.spi = SPI(1)
+        self.spi.init(baudrate=4000_000)
+        self.dc_pin = Pin(DC_PIN, Pin.OUT)
+        
+        self.buffer = bytearray(self.height * self.width // 8)
+        super().__init__(self.buffer, self.width, self.height, framebuf.MONO_HLSB)
+        self.init()
+
+    def digital_write(self, pin, value):
+        pin.value(value)
+
+    def digital_read(self, pin):
+        return pin.value()
+
+    def delay_ms(self, delaytime):
+        utime.sleep(delaytime / 1000.0)
+
+    def spi_writebyte(self, data):
+        self.spi.write(bytearray(data))
+
+    def module_exit(self):
+        self.digital_write(self.reset_pin, 0)
+
+    # Hardware reset
+    def reset(self):
+        self.digital_write(self.reset_pin, 1)
+        self.delay_ms(50) 
+        self.digital_write(self.reset_pin, 0)
+        self.delay_ms(2)
+        self.digital_write(self.reset_pin, 1)
+        self.delay_ms(50)   
+
+    def send_command(self, command):
+        self.digital_write(self.dc_pin, 0)
+        self.digital_write(self.cs_pin, 0)
+        self.spi_writebyte([command])
+        self.digital_write(self.cs_pin, 1)
+
+    def send_data(self, data):
+        self.digital_write(self.dc_pin, 1)
+        self.digital_write(self.cs_pin, 0)
+        self.spi_writebyte([data])
+        self.digital_write(self.cs_pin, 1)
+        
+    def ReadBusy(self):
+        print("e-Paper busy")
+        while(self.digital_read(self.busy_pin) == 1):      #  0: idle, 1: busy
+            self.delay_ms(10) 
+        print("e-Paper busy release")  
+
+    def TurnOnDisplay(self):
+        self.send_command(0x22) # DISPLAY_UPDATE_CONTROL_2
+        self.send_data(0xF7)
+        self.send_command(0x20) # MASTER_ACTIVATION
+        self.ReadBusy()
+
+    def TurnOnDisplay_Partial(self):
+        self.send_command(0x22) # DISPLAY_UPDATE_CONTROL_2
+        self.send_data(0x0F)
+        self.send_command(0x20) # MASTER_ACTIVATION
+        self.ReadBusy()
+
+    def SendLut(self):
+        self.send_command(0x32)
+        for i in range(0, 153):
+            self.send_data(self.lut[i])
+        self.ReadBusy()
+
+    def SetWindow(self, x_start, y_start, x_end, y_end):
+        self.send_command(0x44) # SET_RAM_X_ADDRESS_START_END_POSITION
+        # x point must be the multiple of 8 or the last 3 bits will be ignored
+        self.send_data((x_start>>3) & 0xFF)
+        self.send_data((x_end>>3) & 0xFF)
+        self.send_command(0x45) # SET_RAM_Y_ADDRESS_START_END_POSITION
+        self.send_data(y_start & 0xFF)
+        self.send_data((y_start >> 8) & 0xFF)
+        self.send_data(y_end & 0xFF)
+        self.send_data((y_end >> 8) & 0xFF)
+
+    def SetCursor(self, x, y):
+        self.send_command(0x4E) # SET_RAM_X_ADDRESS_COUNTER
+        self.send_data(x & 0xFF)
+        
+        self.send_command(0x4F) # SET_RAM_Y_ADDRESS_COUNTER
+        self.send_data(y & 0xFF)
+        self.send_data((y >> 8) & 0xFF)
+        self.ReadBusy()
+        
+    def init(self):
+        # EPD hardware init start     
+        self.reset()
+
+        self.ReadBusy()   
+        self.send_command(0x12)  #SWRESET
+        self.ReadBusy()   
+
+        self.send_command(0x01) #Driver output control      
+        self.send_data(0x27)
+        self.send_data(0x01)
+        self.send_data(0x00)
+    
+        self.send_command(0x11) #data entry mode       
+        self.send_data(0x03)
+
+        self.SetWindow(0, 0, self.width-1, self.height-1)
+
+        self.send_command(0x21) #  Display update control
+        self.send_data(0x00)
+        self.send_data(0x80)	
+    
+        self.SetCursor(0, 0)
+        self.ReadBusy()
+        # EPD hardware init end
+        return 0
+
+    def display(self, image):
+        if (image == None):
+            return            
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(0, self.height):
+            for i in range(0, int(self.width / 8)):
+                self.send_data(image[i + j * int(self.width / 8)])   
+        self.TurnOnDisplay()
+
+    def display_Base(self, image):
+        if (image == None):
+            return   
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(0, self.height):
+            for i in range(0, int(self.width / 8)):
+                self.send_data(image[i + j * int(self.width / 8)])
+                
+        self.send_command(0x26) # WRITE_RAM
+        for j in range(0, self.height):
+            for i in range(0, int(self.width / 8)):
+                self.send_data(image[i + j * int(self.width / 8)])   
+                
+        self.TurnOnDisplay()
+        
+    def display_Partial(self, image):
+        if (image == None):
+            return
+            
+        self.digital_write(self.reset_pin, 0)
+        self.delay_ms(2)
+        self.digital_write(self.reset_pin, 1)
+        self.delay_ms(2)   
+        
+        self.SendLut()
+        self.send_command(0x37) 
+        self.send_data(0x00)  
+        self.send_data(0x00)  
+        self.send_data(0x00)  
+        self.send_data(0x00) 
+        self.send_data(0x00)  	
+        self.send_data(0x40)  
+        self.send_data(0x00)  
+        self.send_data(0x00)   
+        self.send_data(0x00)  
+        self.send_data(0x00)
+
+        self.send_command(0x3C) #BorderWavefrom
+        self.send_data(0x80)
+
+        self.send_command(0x22) 
+        self.send_data(0xC0)   
+        self.send_command(0x20) 
+        self.ReadBusy()
+
+        self.SetWindow(0, 0, self.width - 1, self.height - 1)
+        self.SetCursor(0, 0)
+        
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(0, self.height):
+            for i in range(0, int(self.width / 8)):
+                self.send_data(image[i + j * int(self.width / 8)])
+        self.TurnOnDisplay_Partial()
+
+    def Clear(self, color):
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(0, self.height):
+            for i in range(0, int(self.width / 8)):
+                self.send_data(color)
+        self.TurnOnDisplay()
+
+    def sleep(self):
+        self.send_command(0x10) # DEEP_SLEEP_MODE
+        self.send_data(0x01)
+        
+        self.delay_ms(2000)
+        self.module_exit()
+        
+
+class EPD_2in9_Landscape(framebuf.FrameBuffer):
+    def __init__(self):
+        self.reset_pin = Pin(RST_PIN, Pin.OUT)
+        
+        self.busy_pin = Pin(BUSY_PIN, Pin.IN, Pin.PULL_UP)
+        self.cs_pin = Pin(CS_PIN, Pin.OUT)
+        self.width = EPD_WIDTH
+        self.height = EPD_HEIGHT
+        
+        self.lut = WF_PARTIAL_2IN9
+        
+        self.spi = SPI(1)
+        self.spi.init(baudrate=4000_000)
+        self.dc_pin = Pin(DC_PIN, Pin.OUT)
+        
+        self.buffer = bytearray(self.height * self.width // 8)
+        super().__init__(self.buffer, self.height, self.width, framebuf.MONO_VLSB)
+        self.init()
+
+    def digital_write(self, pin, value):
+        pin.value(value)
+
+    def digital_read(self, pin):
+        return pin.value()
+
+    def delay_ms(self, delaytime):
+        utime.sleep(delaytime / 1000.0)
+
+    def spi_writebyte(self, data):
+        self.spi.write(bytearray(data))
+
+    def module_exit(self):
+        self.digital_write(self.reset_pin, 0)
+
+    # Hardware reset
+    def reset(self):
+        self.digital_write(self.reset_pin, 1)
+        self.delay_ms(50) 
+        self.digital_write(self.reset_pin, 0)
+        self.delay_ms(2)
+        self.digital_write(self.reset_pin, 1)
+        self.delay_ms(50)   
+
+    def send_command(self, command):
+        self.digital_write(self.dc_pin, 0)
+        self.digital_write(self.cs_pin, 0)
+        self.spi_writebyte([command])
+        self.digital_write(self.cs_pin, 1)
+
+    def send_data(self, data):
+        self.digital_write(self.dc_pin, 1)
+        self.digital_write(self.cs_pin, 0)
+        self.spi_writebyte([data])
+        self.digital_write(self.cs_pin, 1)
+        
+    def ReadBusy(self):
+        print("e-Paper busy")
+        while(self.digital_read(self.busy_pin) == 1):      #  0: idle, 1: busy
+            self.delay_ms(10) 
+        print("e-Paper busy release")  
+
+    def TurnOnDisplay(self):
+        self.send_command(0x22) # DISPLAY_UPDATE_CONTROL_2
+        self.send_data(0xF7)
+        self.send_command(0x20) # MASTER_ACTIVATION
+        self.ReadBusy()
+
+    def TurnOnDisplay_Partial(self):
+        self.send_command(0x22) # DISPLAY_UPDATE_CONTROL_2
+        self.send_data(0x0F)
+        self.send_command(0x20) # MASTER_ACTIVATION
+        self.ReadBusy()
+
+    def SendLut(self):
+        self.send_command(0x32)
+        for i in range(0, 153):
+            self.send_data(self.lut[i])
+        self.ReadBusy()
+
+    def SetWindow(self, x_start, y_start, x_end, y_end):
+        self.send_command(0x44) # SET_RAM_X_ADDRESS_START_END_POSITION
+        # x point must be the multiple of 8 or the last 3 bits will be ignored
+        self.send_data((x_start>>3) & 0xFF)
+        self.send_data((x_end>>3) & 0xFF)
+        self.send_command(0x45) # SET_RAM_Y_ADDRESS_START_END_POSITION
+        self.send_data(y_start & 0xFF)
+        self.send_data((y_start >> 8) & 0xFF)
+        self.send_data(y_end & 0xFF)
+        self.send_data((y_end >> 8) & 0xFF)
+
+    def SetCursor(self, x, y):
+        self.send_command(0x4E) # SET_RAM_X_ADDRESS_COUNTER
+        self.send_data(x & 0xFF)
+        
+        self.send_command(0x4F) # SET_RAM_Y_ADDRESS_COUNTER
+        self.send_data(y & 0xFF)
+        self.send_data((y >> 8) & 0xFF)
+        self.ReadBusy()
+        
+    def init(self):
+        # EPD hardware init start     
+        self.reset()
+
+        self.ReadBusy()   
+        self.send_command(0x12)  #SWRESET
+        self.ReadBusy()   
+
+        self.send_command(0x01) #Driver output control      
+        self.send_data(0x27)
+        self.send_data(0x01)
+        self.send_data(0x00)
+    
+        self.send_command(0x11) #data entry mode       
+        self.send_data(0x07)
+
+        self.SetWindow(0, 0, self.width-1, self.height-1)
+
+        self.send_command(0x21) #  Display update control
+        self.send_data(0x00)
+        self.send_data(0x80)
+    
+        self.SetCursor(0, 0)
+        self.ReadBusy()
+        # EPD hardware init end
+        return 0
+
+    def display(self, image):
+        if (image == None):
+            return            
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(int(self.width / 8) - 1, -1, -1):
+            for i in range(0, self.height):
+                self.send_data(image[i + j * self.height])   
+        self.TurnOnDisplay()
+
+    def display_Base(self, image):
+        if (image == None):
+            return   
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(int(self.width / 8) - 1, -1, -1):
+            for i in range(0, self.height):
+                self.send_data(image[i + j * self.height])    
+                
+        self.send_command(0x26) # WRITE_RAM
+        for j in range(int(self.width / 8) - 1, -1, -1):
+            for i in range(0, self.height):
+                self.send_data(image[i + j * self.height])      
+                
+        self.TurnOnDisplay()
+        
+    def display_Partial(self, image):
+        if (image == None):
+            return
+            
+        self.digital_write(self.reset_pin, 0)
+        self.delay_ms(2)
+        self.digital_write(self.reset_pin, 1)
+        self.delay_ms(2)   
+        
+        self.SendLut()
+        self.send_command(0x37) 
+        self.send_data(0x00)  
+        self.send_data(0x00)  
+        self.send_data(0x00)  
+        self.send_data(0x00) 
+        self.send_data(0x00)  
+        self.send_data(0x40)  
+        self.send_data(0x00)  
+        self.send_data(0x00)   
+        self.send_data(0x00)  
+        self.send_data(0x00)
+
+        self.send_command(0x3C) #BorderWavefrom
+        self.send_data(0x80)
+
+        self.send_command(0x22) 
+        self.send_data(0xC0)   
+        self.send_command(0x20) 
+        self.ReadBusy()
+
+        self.SetWindow(0, 0, self.width - 1, self.height - 1)
+        self.SetCursor(0, 0)
+        
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(int(self.width / 8) - 1, -1, -1):
+            for i in range(0, self.height):
+                self.send_data(image[i + j * self.height])    
+        self.TurnOnDisplay_Partial()
+
+    def Clear(self, color):
+        self.send_command(0x24) # WRITE_RAM
+        for j in range(int(self.width / 8) - 1, -1, -1):
+            for i in range(0, self.height):
+                self.send_data(color)
+        self.TurnOnDisplay()
+
+    def sleep(self):
+        self.send_command(0x10) # DEEP_SLEEP_MODE
+        self.send_data(0x01)
+        
+        self.delay_ms(2000)
+        self.module_exit()
+
+
+els = None #EPD_2in9_Landscape()
+e = EpaperService(els)
 l = LoggerService()
 s = SettingsService()
-t = TasksService(s, l)
+t = TasksService(s, l, e)
 ws = WifiService(l, s)
+
+
 app = App(l, s, ws, t)
 app.main()
